Fix module Package name
