+++ common-ui/src/main/java/com/jik/common/ui/preview/ThemePreviews.kt
@@ -1,0 +1,8 @@
+package com.jik.common.ui.preview
+
+import android.content.res.Configuration.UI_MODE_NIGHT_YES
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview(name = "dark", group = "themes", uiMode = UI_MODE_NIGHT_YES)
+@Preview(name = "light", group = "themes")
+annotation class ThemePreviews
