--- common-ui/src/main/java/com/jik/common_ui/component/TopAppBar.kt
@@ -1,41 +1,0 @@
-package com.jik.common_ui.component
-
-import androidx.annotation.StringRes
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import com.jik.common_ui.preview.ThemePreviews
-import com.jik.common_ui.theme.MovieTheme
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MovieTopAppBar(
-    @StringRes titleRes: Int,
-    modifier: Modifier = Modifier,
-    colors: TopAppBarColors = TopAppBarDefaults.topAppBarColors(MaterialTheme.colorScheme.primary)
-) {
-    TopAppBar(
-        title = {
-            Text(
-                text = stringResource(id = titleRes),
-                color = MaterialTheme.colorScheme.onPrimary
-            )
-        },
-        colors = colors,
-        modifier = modifier
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@ThemePreviews
-@Composable
-private fun TopAppBarPreview() {
-    MovieTheme(
-        dynamicColor = false
-    ) {
-        MovieTopAppBar(
-            titleRes = android.R.string.untitled
-        )
-    }
-}
