--- common-ui/src/main/java/com/jik/common_ui/preview/ThemePreviews.kt
@@ -1,8 +1,0 @@
-package com.jik.common_ui.preview
-
-import android.content.res.Configuration.UI_MODE_NIGHT_YES
-import androidx.compose.ui.tooling.preview.Preview
-
-@Preview(name = "dark", group = "themes", uiMode = UI_MODE_NIGHT_YES)
-@Preview(name = "light", group = "themes")
-annotation class ThemePreviews
