+++ core-network/src/main/java/com/jik/core/network/source/MovieRemoteDataSourceImpl.kt
@@ -1,0 +1,14 @@
+package com.jik.core.network.source
+
+import com.jik.core.model.Movie
+import com.jik.core.model.MovieInfo
+
+class MovieRemoteDataSourceImpl : MovieRemoteDataSource {
+    override fun getPopularMovies(): List<Movie> {
+        TODO("Not yet implemented")
+    }
+
+    override fun getMovieInfo(id: Int): MovieInfo {
+        TODO("Not yet implemented")
+    }
+}
