--- core-network/src/main/java/com/jik/core_network/source/MovieRemoteDataSource.kt
@@ -1,11 +1,0 @@
-package com.jik.core_network.source
-
-import com.jik.core_model.Movie
-import com.jik.core_model.MovieInfo
-
-interface MovieRemoteDataSource {
-
-    fun getPopularMovies(): List<Movie>
-
-    fun getMovieInfo(id: Int): MovieInfo
-}
