--- core-network/src/main/java/com/jik/core_network/source/MovieRemoteDataSourceImpl.kt
@@ -1,14 +1,0 @@
-package com.jik.core_network.source
-
-import com.jik.core_model.Movie
-import com.jik.core_model.MovieInfo
-
-class MovieRemoteDataSourceImpl : MovieRemoteDataSource {
-    override fun getPopularMovies(): List<Movie> {
-        TODO("Not yet implemented")
-    }
-
-    override fun getMovieInfo(id: Int): MovieInfo {
-        TODO("Not yet implemented")
-    }
-}
